% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exdata}
\alias{exdata}
\title{exdata}
\format{
An object of class \code{data.frame} with 100 rows and 50 columns.
}
\usage{
exdata
}
\description{
Example data of bulk tumor samples (Description)
}
\examples{
  head(exdata)
}
\keyword{datasets}
