% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMAD.R
\name{rMAD}
\alias{rMAD}
\title{Title Evaluation of the tumor heterogeneity level.}
\usage{
rMAD(expr)
}
\arguments{
\item{expr}{gene expression profiles in tumor (microarray or RNA-Seq data, log2 transformed is recommended).}
}
\value{
A dataframe with 2 columns:
\item{Sample}{Tumor samples to be predicted.}
\item{heterogeneity score}{The heterogeneity score of each sample.}
}
\description{
Title Evaluation of the tumor heterogeneity level.
}
\author{
Xiaosheng Wang \href{mailto:xiaosheng.wang@cpu.edu.cn}{xiaosheng.wang@cpu.edu.cn}
}
